+++ src/main/java/ortus/boxlang/modules/orm/ORMContext.java
@@ -204,8 +204,12 @@ public class ORMContext {
 		return this.sessions.computeIfAbsent( sessionKey, ( key ) -> {
 			logger.debug( "opening NEW session for key: {}", sessionKey.getName() );
 
-			SessionFactory sessionFactory = this.ormApp.getSessionFactoryOrThrow( datasource );
-			return sessionFactory.openSession();
+			SessionFactory	sessionFactory	= this.ormApp.getSessionFactoryOrThrow( datasource );
+			Session			session			= sessionFactory.openSession();
+			if ( !config.autoManageSession ) {
+				session.setHibernateFlushMode( org.hibernate.FlushMode.MANUAL );
+			}
+			return session;
 		} );
 	}